Pin the Databricks shuffle manager entries in ProfilingAutoTunerSuiteV2
The two cases added for 14.3 and 17.3 sat in ProfilingAutoTunerSuite, whose header marks it
deprecated and sends new AutoTuner cases to ProfilingAutoTunerSuiteV2. They move there, and a
third case pins 14.3 to 350db143 with the value written out: the map-driven tests take both
the runtime and the expected class from supportedShuffleManagerVersionMap, so a wrong or
missing entry stays green in them, and 17.3 already had its own pin.

verifyRecommendedShuffleManagerVersion and verifyUnsupportedSparkVersionForShuffleManager move
from the deprecated suite into ProfilingAutoTunerSuiteBase, unchanged, so both suites share
them and the map-driven tests keep running.

With the two map entries reverted the three V2 cases fail; with them ProfilingAutoTunerSuite
and ProfilingAutoTunerSuiteV2 pass under JDK 17, 198 succeeded and 0 failed across the two,
scalastyle clean.
